Optimize build of rollout dashboard and include developer documentation.
The developer documentation is now available to users of the dashboard via a new link under the `Resources` menu at the top right of the main screen.
