+++ rollout-dashboard/doc/api.md
@@ -4,9 +4,12 @@
 
 To learn about the API, how to use it, and how to interpret the data
 served by API calls, please consult the programming documentation
-that accompanies the `rollout_dashboard` crate, available under folder
-[`../server/`](../server/) by running the `cargo rustdoc` program within that
-folder, and then launching the Web page it generates for you.
+that accompanies the `rollout_dashboard` crate.  This documentation
+is served by the rollout dashboard UI and is accessible through the
+*Documentation* link at the bottom of the dashboard.  You can also
+build the documentation yourself by running `cargo rustdoc --no-deps --open`
+within the [`../server/`](../server/) folder -- this will launch
+your Web browser with the documentation open.
 
 Please do not proceed with creating a client
 of this server application until you have read that documentation.
