+++ rollout-dashboard/frontend/package.json
@@ -20,14 +20,15 @@
     "flowbite": "^3.1",
     "flowbite-svelte": "^1.13",
     "flowbite-svelte-icons": "^3",
+    "rollup-plugin-visualizer": "^6.0.3",
     "svelte": "^5",
     "svelte-check": "^4",
     "svelte-copy": "^2",
     "svelte-time": "^2",
-    "tailwindcss": "^4",
     "tailwind-variants": "^3",
+    "tailwindcss": "^4",
     "tslib": "^2",
     "typescript": "^5",
-    "vite": "^7"
+    "vite": "^6"
   }
 }
