+++ rollout-dashboard/frontend/tsconfig.json
@@ -1,22 +1,13 @@
 {
   "extends": "@tsconfig/svelte/tsconfig.json",
   "compilerOptions": {
-    "target": "ESNext",
+    "module": "esnext",
+    "target": "es2018",
     "useDefineForClassFields": true,
-    "module": "ESNext",
     "resolveJsonModule": true,
-    /**
-     * Typecheck JS in `.svelte` and `.js` files by default.
-     * Disable checkJs if you'd like to use dynamic types in JS.
-     * Note that setting allowJs false does not prevent the use
-     * of JS in `.svelte` files.
-     */
-    "verbatimModuleSyntax": true,
     "allowJs": true,
-    "sourceMap": true,
     "checkJs": true,
     "isolatedModules": true,
-    "moduleDetection": "force",
     "outDir": "./dist"
   },
   "include": [
