+++ rollout-dashboard/frontend/tsconfig.node.json
@@ -3,7 +3,8 @@
     "composite": true,
     "tsBuildInfoFile": "./node_modules/.tmp/tsconfig.node.tsbuildinfo",
     "skipLibCheck": true,
-    "module": "ESNext",
+    "module": "esnext",
+    "target": "es2018",
     "moduleResolution": "bundler",
     "strict": true,
   },
