+++ rollout-dashboard/server/src/lib.rs
@@ -1,25 +1,33 @@
-//! Rollout dashboard interoperability library.
-//!
 //! The rollout dashboard is a service that provides up-to-date
-//! access to rollout information.  This service provides a REST
-//! endpoint to obtain that information under `/api/v1`.  The
-//! endpoint is unauthenticated and data returned to the client
-//! is always encoded as JSON using the serde serialization library.
+//! access to rollout information.  It is composed of two parts:
+//!
+//! * The frontend (the user interface that probably led you here).
+//! * The backend server (used by the frontend to fetch data).
+//!
+//! ## Using the backend server API
+//!
+//! The backend server provides several versioned endpoints under
+//! `/api`.  The endpoints are unauthenticated.
+//! Data returned to the client from non-SSE requests is always encoded
+//! as JSON using the serde serialization library.  When SSE requests
+//! are sent, responses are in SSE events form with JSON-encoded objects
+//! as payload.
+//!
+//! Consult the [api_server] module to learn about the request endpoints
+//! supported by the API.  See the [types] library to understand the
+//! various responses that the API provides.
+//!
 //! Clients are encouraged to use the types available in this library
 //! to deserialize responses from the API calls.
 //!
-//! The only REST call that exists today is `GET /api/v1/rollouts`.
+//! ## Learn more about the rollout dashboard
 //!
-//! * When successful, it returns a [`types::Rollouts`] instance
-//!   within an HTTP 200 response code.
-//! * When there are no rollouts available yet (because the service
-//!   has just begun loading), an HTTP 204 No Content status is returned.
-//! * In case of errors contacting Airflow or processing its data, it will
-//!   return an HTTP status code and an appropriate error message.
-//!
-//! The call also exists in a server-side events streaming flavor,
-//! available by appending `/sse` to the rollouts API call.  The server
-//! will stream the initial state right away, then stream updates to
-//! the rollout data as they happen.
+//! To learn more about the rollout dashboard, please consult the
+//! [README](https://github.com/dfinity/dre-airflow/tree/main/rollout-dashboard)
+//! of the project.
+#[doc(hidden)]
 pub mod airflow_client;
+pub mod api_server;
+#[doc(hidden)]
+pub mod live_state;
 pub mod types;
